requirements.txt
Cython==3.0.11
cython_bbox==0.1.5
easydict==1.13
einops==0.8.0
huggingface-hub==0.16.4
imagecorruptions==1.1.2
imageio==2.31.2
lap==0.4.0
matplotlib==3.5.3
motmetrics==1.4.0
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
opencv-python==4.10.0.84
pandas==2.0.3
pillow==10.4.0
scikit-image==0.19.3
scikit-learn==0.22.2
scipy==1.10.1
timm==1.0.11
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.67.0